rename new test files to avoid conflicts
